refactor(titles): lift backend probing out of the kernel module
Refs #685 — the `titles/taichi_kernels.py` entry in the split wave.

taichi_kernels.py was 896 lines carrying two unrelated jobs: deciding
which Taichi arch can dispatch on this machine, and defining the title
kernels themselves. The first knows nothing about the second — it owns
the taichi import guard, an fd-level stdout silencer, and a spawn-child
probe that proves a backend before the parent commits to it. It already
had its own test file.

The kernels stay put, and that is the point. Consumers reach them as
module attributes at call time (`taichi_kernels._render_sdf_text`),
precisely so `init_taichi()` can compile them after import. Splitting the
kernels across modules would have meant a `_compile_*` function, a set of
None globals and an init hook per module — the compile wiring is what
couples them, so the honest seam is elsewhere.

  taichi_kernels.py       896 -> 722
  taichi_backend_probe.py   - -> 207

The env guard keeps its ordering: the new module sets
ENABLE_TAICHI_HEADER_PRINT/TI_LOG_LEVEL beside its `import taichi`, and
taichi_kernels imports it above the SDF block, which pulls in Taichi too.

Patch-target audit — `init_taichi` spans both namespaces now, so the
existing test was repointed name by name rather than by search-replace:
TAICHI_AVAILABLE, the init flags, `_silent_init` and `_compile_kernels`
still resolve in `taichi_kernels` and stay patched there; `ti` and
`_probe_taichi_backend` are read by `_candidate_backends` and
`_backend_dispatches`, so those two move to the probe module. Patching
either one on the wrong module would have silently left the real Taichi
arch objects in play.

`_taichi_probe_worker` had no test — it only ever ran in a spawned child.
Four added tests drive it with a fake Taichi module, covering the case
the probe exists for: init() succeeding while nothing is dispatched.
Module coverage 76% -> 90%.
