+++ pyproject.toml
@@ -454,6 +454,7 @@ module = [
     "immich_memories.titles.renderer_taichi",
     "immich_memories.titles.taichi_text",
     "immich_memories.titles.sdf_font",
+    "immich_memories.titles.taichi_backend_probe",
     "immich_memories.titles.taichi_kernels",
     "immich_memories.titles.sdf_font_rendering",
     "immich_memories.titles.taichi_video",
