+++ src/immich_memories/titles/taichi_kernels.py
@@ -1,28 +1,28 @@
 """Taichi GPU kernel definitions, initialization, and helper functions.
 
+Backend selection lives in taichi_backend_probe.py; this module owns the
+kernels themselves and the lazy compilation that init_taichi() drives.
+
 Note: This module does NOT use 'from __future__ import annotations'
 because Taichi kernels require actual type objects, not string annotations.
 """
 
 import contextlib
 import logging
-import os
-import queue
-import sys
-from dataclasses import dataclass
-from enum import StrEnum
 from pathlib import Path
 
 import numpy as np
 
-logger = logging.getLogger(__name__)
-
-# WHY: Taichi's C++ runtime prints to stdout, corrupting Rich Live display.
-# ENABLE_TAICHI_HEADER_PRINT="0" — suppresses import-time version banner (taichi#8334)
-# TI_LOG_LEVEL — belt-and-suspenders for C++ log messages
-# The main fix is verbose=False on ti.init() calls (see init_taichi and _check_taichi)
-os.environ.setdefault("ENABLE_TAICHI_HEADER_PRINT", "0")
-os.environ.setdefault("TI_LOG_LEVEL", "error")
+# WHY: importing this first is what sets ENABLE_TAICHI_HEADER_PRINT/TI_LOG_LEVEL
+# before any module in this package pulls in Taichi's C++ runtime — including
+# the SDF modules below, which import Taichi themselves.
+from .taichi_backend_probe import (
+    TAICHI_AVAILABLE,
+    _backend_dispatches,
+    _candidate_backends,
+    _silent_init,
+    ti,
+)
 
 try:
     from .sdf_atlas_gen import get_cached_atlas
@@ -37,185 +37,11 @@ except ImportError:
     init_sdf_kernels = None
     layout_text = None
 
-try:
-    import taichi as ti
-
-    TAICHI_AVAILABLE = True
-except ImportError:
-    TAICHI_AVAILABLE = False
-    ti = None
+logger = logging.getLogger(__name__)
 
 _taichi_initialized = False
 _taichi_backend = None
 _kernels_compiled = False
-
-_TAICHI_PROBE_TIMEOUT_SECONDS = 10.0
-_TAICHI_PROBE_STOP_TIMEOUT_SECONDS = 1.0
-
-
-class TaichiProbeOutcome(StrEnum):
-    """Bounded outcomes from an isolated backend dispatch probe."""
-
-    SUCCESS = "success"
-    DISPATCH_FAILED = "dispatch_failed"
-    CHILD_CRASHED = "child_crashed"
-    TIMED_OUT = "timed_out"
-
-
-@dataclass(frozen=True)
-class TaichiProbeResult:
-    """Small picklable result sent from the probe child to its parent."""
-
-    outcome: TaichiProbeOutcome
-    detail: str | None = None
-
-
-@contextlib.contextmanager
-def _silence_output_fds():
-    """Temporarily redirect process stdout/stderr to the null device."""
-    sys.stdout.flush()
-    sys.stderr.flush()
-    devnull_fd = os.open(os.devnull, os.O_WRONLY)
-    saved_stdout = os.dup(1)
-    saved_stderr = os.dup(2)
-    try:
-        os.dup2(devnull_fd, 1)
-        os.dup2(devnull_fd, 2)
-        yield
-    finally:
-        sys.stdout.flush()
-        sys.stderr.flush()
-        os.dup2(saved_stdout, 1)
-        os.dup2(saved_stderr, 2)
-        os.close(saved_stdout)
-        os.close(saved_stderr)
-        os.close(devnull_fd)
-
-
-def _silent_init(**kwargs) -> None:
-    """Call ti.init() with stdout/stderr silenced at the OS file descriptor level.
-
-    WHY: Taichi's C++ runtime prints "[Taichi] Starting on arch=metal"
-    directly to file descriptor 1, bypassing Python's sys.stdout, all
-    env vars (TI_LOG_LEVEL, ENABLE_TAICHI_HEADER_PRINT), and all Python
-    API flags (verbose=False, log_level). The ONLY way to suppress it is
-    to redirect the raw OS file descriptors during the call.
-    """
-    with _silence_output_fds():
-        ti.init(**kwargs)
-
-
-def _taichi_probe_worker(backend_name: str, result_queue) -> None:
-    """Initialize one backend and dispatch a real kernel inside a child process."""
-    try:
-        with _silence_output_fds():
-            backend = getattr(ti, backend_name)
-            ti.init(arch=backend, offline_cache=True)
-
-            @ti.kernel
-            def increment(values: ti.types.ndarray(dtype=ti.i32, ndim=1)):
-                for index in values:
-                    values[index] += 1
-
-            values = np.zeros(1, dtype=np.int32)
-            increment(values)
-        if values[0] != 1:
-            result = TaichiProbeResult(
-                TaichiProbeOutcome.DISPATCH_FAILED,
-                "unexpected_kernel_result",
-            )
-        else:
-            result = TaichiProbeResult(TaichiProbeOutcome.SUCCESS)
-    except Exception as exc:
-        result = TaichiProbeResult(
-            TaichiProbeOutcome.DISPATCH_FAILED,
-            type(exc).__name__,
-        )
-    result_queue.put(result)
-
-
-def _stop_probe_process(process) -> None:
-    """Stop a stuck probe, escalating from terminate to kill."""
-    process.terminate()
-    process.join(_TAICHI_PROBE_STOP_TIMEOUT_SECONDS)
-    if process.is_alive():
-        process.kill()
-        process.join(_TAICHI_PROBE_STOP_TIMEOUT_SECONDS)
-
-
-def _probe_taichi_backend(
-    backend_name: str,
-    timeout: float = _TAICHI_PROBE_TIMEOUT_SECONDS,
-) -> TaichiProbeResult:
-    """Probe one non-CPU backend without changing parent Taichi state."""
-    import multiprocessing
-
-    context = multiprocessing.get_context("spawn")
-    result_queue = context.Queue(maxsize=1)
-    process = context.Process(
-        target=_taichi_probe_worker,
-        args=(backend_name, result_queue),
-        name=f"taichi-{backend_name}-probe",
-        daemon=True,
-    )
-    started = False
-    try:
-        process.start()
-        started = True
-        process.join(timeout)
-        if process.is_alive():
-            _stop_probe_process(process)
-            return TaichiProbeResult(TaichiProbeOutcome.TIMED_OUT)
-        try:
-            result = result_queue.get(timeout=0.25)
-        except queue.Empty:
-            return TaichiProbeResult(
-                TaichiProbeOutcome.CHILD_CRASHED,
-                f"exitcode={process.exitcode}",
-            )
-        if not isinstance(result, TaichiProbeResult):
-            return TaichiProbeResult(TaichiProbeOutcome.CHILD_CRASHED, "invalid_result")
-        return result
-    except (OSError, RuntimeError, TypeError) as exc:
-        return TaichiProbeResult(TaichiProbeOutcome.CHILD_CRASHED, type(exc).__name__)
-    finally:
-        if started and process.is_alive():
-            _stop_probe_process(process)
-        result_queue.close()
-        result_queue.join_thread()
-        if started and not process.is_alive():
-            process.close()
-
-
-def _candidate_backends(
-    *, force_cpu: bool, operating_system: str
-) -> list[tuple[object, str, str | None]]:
-    """Return parent architecture objects and child-safe probe names in priority order."""
-    if force_cpu:
-        return [(ti.cpu, "CPU", None)]
-    if operating_system == "Darwin":
-        return [(ti.metal, "Metal", "metal"), (ti.cpu, "CPU", None)]
-    return [
-        (ti.cuda, "CUDA", "cuda"),
-        (ti.vulkan, "Vulkan", "vulkan"),
-        (ti.cpu, "CPU", None),
-    ]
-
-
-def _backend_dispatches(name: str, probe_name: str | None) -> bool:
-    """Prove a GPU backend can dispatch, while allowing CPU to bypass the probe."""
-    if probe_name is None:
-        return True
-    probe = _probe_taichi_backend(probe_name)
-    if probe.outcome is TaichiProbeOutcome.SUCCESS:
-        return True
-    logger.debug(
-        "Taichi %s dispatch probe failed (%s: %s)",
-        name,
-        probe.outcome.value,
-        probe.detail or "no detail",
-    )
-    return False
 
 
 def init_taichi() -> str | None:
